refactor(packages/supabase): add returning type
