+++ packages/supabase/src/client/client.ts
@@ -1,5 +1,6 @@
 import { createBrowserClient } from "@supabase/ssr";
-import type { Database } from "../types";
+import { SupabaseClient } from "@supabase/supabase-js";
+import { Database } from "../types/database";
 
 if (!process.env.NEXT_PUBLIC_SUPABASE_URL) {
   throw new Error("Missing env.NEXT_PUBLIC_SUPABASE_URL");
@@ -9,7 +10,7 @@ if (!process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY) {
   throw new Error("Missing env.NEXT_PUBLIC_SUPABASE_ANON_KEY");
 }
 
-export function createSupabaseClient() {
+export function createSupabaseClient(): SupabaseClient<Database> {
   return createBrowserClient<Database>(
     process.env.NEXT_PUBLIC_SUPABASE_URL!,
     process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
