+++ packages/supabase/src/client/server.ts
@@ -1,6 +1,7 @@
 import { createServerClient } from "@supabase/ssr";
-import type { Database } from "../types";
+import { Database } from "../types/database";
 import { cookies } from "next/headers";
+import { SupabaseClient } from "@supabase/supabase-js";
 
 if (!process.env.NEXT_PUBLIC_SUPABASE_URL) {
   throw new Error("Missing env.NEXT_PUBLIC_SUPABASE_URL");
@@ -10,7 +11,7 @@ if (!process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY) {
   throw new Error("Missing env.NEXT_PUBLIC_SUPABASE_ANON_KEY");
 }
 
-export function createSupabaseServer() {
+export function createSupabaseServer(): SupabaseClient<Database> {
   const cookieStore = cookies();
 
   return createServerClient<Database>(
